+++ src/dune_rpc_impl/import.ml
@@ -2,3 +2,5 @@ include Stdune
 module Console = Dune_console
 module Dune_rpc = Dune_rpc_private
 module Build_system = Dune_engine.Build_system
+module Where = Dune_rpc_client.Where
+module Client = Dune_rpc_client.Client