+++ test/blackbox-tests/test-cases/melange/emit-private.t
@@ -24,13 +24,13 @@ Test dependency on a private library in the same package as melange.emit
   $ dune install --prefix $PWD/prefix
   Installing $TESTCASE_ROOT/prefix/lib/a/META
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a.cmi
-  Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a.cmj
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a.cmt
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a__Foo.cmi
-  Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a__Foo.cmj
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/.public_cmi_melange/a__Foo.cmt
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/a.ml
   Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/foo.ml
+  Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/melange/a.cmj
+  Installing $TESTCASE_ROOT/prefix/lib/a/__private__/a/melange/a__Foo.cmj
   Installing $TESTCASE_ROOT/prefix/lib/a/dune-package
 
   $ cat > b/dune <<EOF